Analytics outliers API: period not fully supported. Missing fixed types.[DHIS2-16894]
