+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/outlier/data/OutlierQueryParams.java
@@ -38,7 +38,6 @@ import org.hisp.dhis.analytics.OutlierDetectionAlgorithm;
 import org.hisp.dhis.analytics.QueryKey;
 import org.hisp.dhis.analytics.SortOrder;
 import org.hisp.dhis.common.IdScheme;
-import org.hisp.dhis.period.RelativePeriodEnum;
 
 /** Encapsulation of a web API request for outlier value detection. */
 @Data
@@ -49,7 +48,7 @@ public class OutlierQueryParams {
 
   private Set<String> ou = new HashSet<>();
 
-  private RelativePeriodEnum pe;
+  private String pe;
 
   /**
    * This parameter selects the headers to be returned in the response. We use a LinkedHashSet
