+++ dhis-2/dhis-test-e2e/src/test/java/org/hisp/dhis/analytics/outlier/OutliersDetection4AutoTest.java
@@ -962,157 +962,4 @@ public class OutliersDetection4AutoTest extends AnalyticsApiTest {
             "3.6560656799",
             "-294.6529422297"));
   }
-
-  @Test
-  public void queryOutliertest17() throws JSONException {
-    // Given
-    QueryParamsBuilder params =
-        new QueryParamsBuilder()
-            .add("endDate=2022-10-26")
-            .add("outputIdScheme=uid")
-            .add("sortOrder=asc")
-            .add("orderBy=zscore")
-            .add("threshold=3")
-            .add("startDate=2022-07-26")
-            .add("ds=BfMAe6Itzgt")
-            .add("algorithm=Z_SCORE")
-            .add("skipRounding=true")
-            .add("relativePeriodDate=2022-07-01");
-
-    // When
-    ApiResponse response = actions.query().get("", JSON, JSON, params);
-
-    // Then
-    response
-        .validate()
-        .statusCode(200)
-        .body("headers", hasSize(equalTo(18)))
-        .body("rows", hasSize(equalTo(3)))
-        .body("height", equalTo(3))
-        .body("width", equalTo(18))
-        .body("headerWidth", equalTo(18));
-
-    // Assert metaData.
-    String expectedMetaData =
-        "{\"count\":3,\"orderBy\":\"Z_SCORE\",\"threshold\":\"3.0\",\"maxResults\":500,\"algorithm\":\"Z_SCORE\"}";
-    String actualMetaData = new JSONObject((Map) response.extract("metaData")).toString();
-    assertEquals(expectedMetaData, actualMetaData, false);
-
-    // Assert headers.
-    validateHeader(response, 0, "dx", "Data", "TEXT", "java.lang.String", false, false);
-    validateHeader(response, 1, "dxname", "Data name", "TEXT", "java.lang.String", false, false);
-    validateHeader(response, 2, "pe", "Period", "TEXT", "java.lang.String", false, false);
-    validateHeader(response, 3, "pename", "Period name", "TEXT", "java.lang.String", false, false);
-    validateHeader(
-        response, 4, "ou", "Organisation unit", "TEXT", "java.lang.String", false, false);
-    validateHeader(
-        response, 5, "ouname", "Organisation unit name", "TEXT", "java.lang.String", false, false);
-    validateHeader(
-        response,
-        6,
-        "ounamehierarchy",
-        "Organisation unit name hierarchy",
-        "TEXT",
-        "java.lang.String",
-        false,
-        false);
-    validateHeader(
-        response, 7, "coc", "Category option combo", "TEXT", "java.lang.String", false, false);
-    validateHeader(
-        response,
-        8,
-        "cocname",
-        "Category option combo name",
-        "TEXT",
-        "java.lang.String",
-        false,
-        false);
-    validateHeader(
-        response, 9, "aoc", "Attribute option combo", "TEXT", "java.lang.String", false, false);
-    validateHeader(
-        response,
-        10,
-        "aocname",
-        "Attribute option combo name",
-        "TEXT",
-        "java.lang.String",
-        false,
-        false);
-    validateHeader(response, 11, "value", "Value", "NUMBER", "java.lang.Double", false, false);
-    validateHeader(response, 12, "mean", "Mean", "NUMBER", "java.lang.Double", false, false);
-    validateHeader(
-        response, 13, "stddev", "Standard deviation", "NUMBER", "java.lang.Double", false, false);
-    validateHeader(
-        response, 14, "absdev", "Absolute deviation", "NUMBER", "java.lang.Double", false, false);
-    validateHeader(response, 15, "zscore", "Z-score", "NUMBER", "java.lang.Double", false, false);
-    validateHeader(
-        response, 16, "lowerbound", "Lower boundary", "NUMBER", "java.lang.Double", false, false);
-    validateHeader(
-        response, 17, "upperbound", "Upper boundary", "NUMBER", "java.lang.Double", false, false);
-
-    // Assert rows.
-    validateRow(
-        response,
-        List.of(
-            "dU0GquGkGQr",
-            "Q_Early breastfeeding (within 1 hr after delivery) at BCG",
-            "202209",
-            "September 2022",
-            "Mi4dWRtfIOC",
-            "Sandaru CHC",
-            "Sierra Leone / Kailahun / Penguia / Sandaru CHC",
-            "Prlt0C1RF0s",
-            "Fixed, <1y",
-            "HllvX50cXC0",
-            "default",
-            "105.0",
-            "18.2635658915",
-            "28.7055404478",
-            "86.7364341085",
-            "3.0215920953",
-            "-67.8530554519",
-            "104.3801872349"));
-    validateRow(
-        response,
-        List.of(
-            "s46m5MS0hxu",
-            "BCG doses given",
-            "202208",
-            "August 2022",
-            "CvBAqD6RzLZ",
-            "Ngalu CHC",
-            "Sierra Leone / Bo / Bargbe / Ngalu CHC",
-            "Prlt0C1RF0s",
-            "Fixed, <1y",
-            "HllvX50cXC0",
-            "default",
-            "220.0",
-            "41.6440677966",
-            "57.4495394538",
-            "178.3559322034",
-            "3.1045667885",
-            "-130.7045505648",
-            "213.992686158"));
-    validateRow(
-        response,
-        List.of(
-            "l6byfWFUGaP",
-            "Yellow Fever doses given",
-            "202209",
-            "September 2022",
-            "RhJbg8UD75Q",
-            "Yemoh Town CHC",
-            "Sierra Leone / Bo / Kakua / Yemoh Town CHC",
-            "Prlt0C1RF0s",
-            "Fixed, <1y",
-            "HllvX50cXC0",
-            "default",
-            "466.0",
-            "48.1860465116",
-            "114.2796629138",
-            "417.8139534884",
-            "3.6560656799",
-            "-294.6529422297",
-            "391.025035253"));
-  }
 }
